reworked circular issue
